Time: 0 ms (100.00%), Space: 7 MB (98.74%) - LeetHub

+++ 1491-average-salary-excluding-the-minimum-and-maximum-salary/1491-average-salary-excluding-the-minimum-and-maximum-salary.cpp
@@ -1,12 +1,12 @@
 class Solution {
 public:
     double average(vector<int>& salary) {
-      float mini=1e6;
-      float maxi=1000;
+      double  mini=1e6;
+      double maxi=1000;
         double sum=0;
         double cnt=0;
         for(int i=0; i<salary.size();i++){
-            float x=salary[i];
+            double x=salary[i];
             mini=min(mini,x);
             maxi=max(maxi,x);
         }
